avplay app: split AvReader file to hpp/cpp
